+++ app/src/main/java/com/poti/android/domain/model/HistorySummary.kt
@@ -1,0 +1,10 @@
+package com.poti.android.domain.model
+
+import androidx.annotation.StringRes
+import com.poti.android.presentation.user.component.HistorySummaryType
+
+data class HistorySummaryItem(
+    val type: HistorySummaryType,
+    @StringRes val titleRes: Int,
+    val count: Int,
+)
